Adaptive_Rag/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
tiktoken
pinecone-client
python-dotenv
openai
langchain_tavily
langchain_pinecone